+++ package.json
@@ -11,9 +11,9 @@
   "module": "./index.js",
   "types": "./dist/index.d.ts",
   "exports": {
-    "./url-pattern": {
-      "import": "./dist/url-pattern.js",
-      "require": "./dist/url-pattern.cjs"
+    "./urlpattern": {
+      "import": "./dist/urlpattern.js",
+      "require": "./dist/urlpattern.cjs"
     },
     ".": {
       "import": "./index.js",
@@ -54,8 +54,8 @@
   },
   "scripts": {
     "prebuild": "rimraf dist",
-    "build:esm": "esbuild --bundle --format=esm src/url-pattern.ts --outfile=dist/url-pattern.js",
-    "build:cjs": "esbuild --bundle --format=cjs src/url-pattern.ts --outfile=dist/url-pattern.cjs",
+    "build:esm": "esbuild --bundle --format=esm src/url-pattern.ts --outfile=dist/urlpattern.js",
+    "build:cjs": "esbuild --bundle --format=cjs src/url-pattern.ts --outfile=dist/urlpattern.cjs",
     "copy:dts": "cp ./src/index.d.ts ./src/types.d.ts ./dist",
     "build": "npm run build:esm && npm run build:cjs && npm run copy:dts",
     "pretest": "npm run build; rm -rf node_modules/urlpattern-polyfill; ln -s $(pwd) node_modules/urlpattern-polyfill",
