--- url-pattern/package.json
@@ -1,5 +1,0 @@
-{
-  "main": "../dist/url-pattern.cjs",
-  "module": "../dist/url-pattern.js",
-  "types": "../dist/types.d.ts"
-}
